+++ mvc/src/main/java/nextstep/mvc/servlet/DispatcherServlet.java
@@ -4,8 +4,8 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.HandlerMapping;
+import nextstep.mvc.handleradapter.HandlerAdapter;
+import nextstep.mvc.controller.tobe.HandlerMapping;
 import nextstep.mvc.view.ModelAndView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
