+++ mvc/src/main/java/nextstep/mvc/servlet/HandlerMappingRegistry.java
@@ -4,7 +4,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import nextstep.mvc.HandlerMapping;
+import nextstep.mvc.controller.tobe.HandlerMapping;
 
 public class HandlerMappingRegistry {
